poc/requirements.txt
# AEGIS POC - Python Dependencies
# Core cryptography
pyjwt[crypto]==2.8.0
cryptography==42.0.0

# Utilities
python-dateutil==2.8.2

# Development/Testing
pytest==7.4.3
pytest-cov==4.1.0
